# spec/dashboard/games_openapi.yaml
openapi: 3.0.0
info:
  description: "Qilin api "
  version: "1.0.0"
  title: "Qilin games dashboard API"
  contact:
    email: "roman.golenok@protocol.one"
...
servers:
  - url: https://qilinapi.tst.protocol.one
    description: Master test server
paths:
...
  /api/v1/games/{id}:
    get:
      tags:
        - "game"
      summary: "Get game info"
      description: ""
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          required: true
          description: "Id of game"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameInfo"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'


    delete:
      tags:
        - "game"
      summary: "Delete game"
      description: ""
      operationId: "deleteGame"
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          required: true
          description: "Id of game to remove"
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - "game"
      summary: "Change game info"
      description: ""
      operationId: "changeGame"
      requestBody:
        description: "Game object that needs to be saved in store"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeGame"
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
            format: uuid
          required: true
          description: "Id of game to change"
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: false
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'
...
  /api/v1/games/{id}/media:
    get:
      tags:
        - media
      summary: "Gets media for game"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
    put:
      tags:
        - media
      summary: "Change media for game"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Media'
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal Server Error
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 409
            message: "Resource with same unique id already exists"
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 422
            message: "Fields required: 'title, description'"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Media:
      type: object
      properties:
        coverImage:
          type: object
          additionalProperties:
            type: string
            format: uri
          example:
            en: http://www.example.com/images/ce05bce1-48bf-4c01-ae14-046af4d1ca94
            ru: http://www.example.com/images/30083025-fe0c-4a22-b0a7-569ab6cee50b
        coverVideo:
          type: object
          additionalProperties:
            type: string
            format: uri
          example:
            en: http://www.example.com/video/ce05bce1-48bf-4c01-ae14-046af4d1ca94
            ru: http://www.example.com/video/30083025-fe0c-4a22-b0a7-569ab6cee50b
        trailers:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
              format: uri
          example:
            en:
              - http://www.example.com/video/ce05bce1-48bf-4c01-ae14-046af4d1ca94
              - http://www.example.com/video/30083025-fe0c-4a22-b0a7-569ab6cee50b
            ru:
              - http://www.example.com/video/ce05bce1-48bf-4c01-ae14-046af4d1ca94
              - http://www.example.com/video/30083025-fe0c-4a22-b0a7-569ab6cee50b
        screenshots:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
              format: uri
          example:
            en:
              - http://www.example.com/video/ce05bce1-48bf-4c01-ae14-046af4d1ca94
              - http://www.example.com/video/30083025-fe0c-4a22-b0a7-569ab6cee50b
            ru:
              - http://www.example.com/video/ce05bce1-48bf-4c01-ae14-046af4d1ca94
              - http://www.example.com/video/30083025-fe0c-4a22-b0a7-569ab6cee50b
        store:
          type: object
          properties:
            special:
              type: object
              additionalProperties:
                type: string
                format: uri
              example:
                en: http://www.example.com/images/ce05bce1-48bf-4c01-ae14-046af4d1ca94
                ru: http://www.example.com/images/30083025-fe0c-4a22-b0a7-569ab6cee50b
            friends:
              type: object
              additionalProperties:
                type: string
                format: uri
              example:
                en: http://www.example.com/images/ce05bce1-48bf-4c01-ae14-046af4d1ca94
                ru: http://www.example.com/images/30083025-fe0c-4a22-b0a7-569ab6cee50b
        capsule:
          type: object
          properties:
            generic:
              additionalProperties:
                type: string
                format: uri
              example:
                en: http://www.example.com/images/ce05bce1-48bf-4c01-ae14-046af4d1ca94
                ru: http://www.example.com/images/30083025-fe0c-4a22-b0a7-569ab6cee50b
            small:
              additionalProperties:
                type: string
                format: uri
              example:
                en: http://www.example.com/images/ce05bce1-48bf-4c01-ae14-046af4d1ca94
                ru: http://www.example.com/images/30083025-fe0c-4a22-b0a7-569ab6cee50b
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    LanguageSupport:
      type: object
      properties:
        voice:
          type: "boolean"
          default: false
        interface:
          type: "boolean"
          default: false
        subtitles:
          type: "boolean"
          default: false
    Languages:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/LanguageSupport"
      example:
        en:
          voice: true
          interface: true
          subtitles: true
        ru:
          voice: false
          interface: true
          subtitles: true
        fr:
          voice: false
          interface: false
          subtitles: true
...
    GameGenres:
      type: object
      properties:
        main:
          type: integer
        addition:
          type: array
          items:
            type: integer
...
    Features:
      type: object
      properties:
        common:
          type: array
          items:
            type: string
            enum:
              - singlePlayer
              - multiPlayer
              - onlineMultiplayerPlayer
              - localMultiplayerPlayer
              - splitScreen
              - mmo
              - cooperative
              - onlineCooperative
              - localCooperative
              - crossPlatform
              - achievements
              - cloudSaves
        controllers:
          type: "string"
          enum:
            - "not_supported"
            - "partially_supported"
            - "full_supported"
          default: "not_supported"
    SupportPlatform:
      type: object
      properties:
        windows:
          type: "boolean"
          default: true
        macOs:
          type: "boolean"
          default: false
        linux:
          type: "boolean"
          default: false
    MachineRequirements:
      type: object
      properties:
        system:
          type: "string"
          example: "Windows"
        processor:
          type: "string"
          example: "Intel i9"
        graphics:
          type: "string"
          example: "NVIDIA GeForce RTX2080Ti or similar"
        sound:
          type: "string"
        ram:
          type: "integer"
          example: 4
        ramDimension:
          type: string
          enum:
            - MB
            - GB
        storage:
          type: "integer"
        storageDimension:
          type: string
          enum:
            - MB
            - GB
        other:
          type: "string"
...
    BaseGame:
      type: object
      properties:
        internalName:
          type: "string"
          readOnly: true
        # non-localized title of game for end user
        title:
          type: "string"
        # comma separated list of developers
        developers:
          type: "string"
        # comma separated list of publishers
        publishers:
          type: "string"
        releaseDate:
          type: "string"
          format: "date-time"
        displayRemainingTime:
          type: "boolean"
          default: false
        features:
          $ref: "#/components/schemas/Features"
        platforms:
          $ref: "#/components/schemas/SupportPlatform"
        requirements:
          type: object
          properties:
            macOs:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
            linux:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
            windows:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
        languages:
          $ref: "#/components/schemas/Languages"
        achievementOnProd:
          type: boolean
          default: false
          description: "Enables achievements on production"
    GameInfo:
      allOf:
        - $ref: '#/components/schemas/BaseGame'
        - type: object
          properties:
            genres:
              $ref: '#/components/schemas/GameGenres'
            tags:
              type: array
              items:
                type: integer
    ChangeGame:
      allOf:
        - $ref: '#/components/schemas/BaseGame'
        - type: object
          properties:
            genres:
              $ref: '#/components/schemas/GameGenres'
            tags:
              type: array
              items:
                type: integer
...
security:
  - bearerAuth: []